chore(webhooks): move `triggerEvent` to `server.ts`
